ENHANCE: Move project back into surround lib
